+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add `/whoami` API endpoint for improved supportability and debugging for access
   tokens and client IP address determination. [cyberark/conjur#1697](https://github.com/cyberark/conjur/issues/1697)
+- `TRUSTED_PROXIES` is validated at Conjur startup to ensure that it contains
+  valid IP addresses and/or address ranges in CIDR notation.
+  [cyberark/conjur#1727](https://github.com/cyberark/conjur/issues/1727)
 
 ### Changed
 - The Conjur server request logs now records the same IP address used by audit